fix: overflow in ordinary redistributions (#396)
This PR fixes an overflow issue with ordinary redistributions when the 
recipient yang pool is a very small value by requiring a minimum
amount of 1 Wad. Otherwise, exception redistribution occurs.

It also includes the initial yang amount when calculating the unit yang
to slightly improve the precision of both the redistributed value and 
the remainder value in the redistributed trove.