fix: revert if zero shares when providing; increase initial shares (#543)
This PR changes the behaviour of `absorber.provide` to revert if the resulting amount of 
shares is zero, so as to prevent users from inadvertently donating their yin. 

In addition the initial shares has been increased from 10**3 to 10**9, which would now 
require a donation of 10**9 Wad after the initial shares in order for a provision of 1 Wad 
to revert due to zero shares, thereby mitigating against potential griefing.