script: simulate absorption (#580)
This PR adds a script that sets up a devnet instance with an absorbable 
trove for the purpose of testing the keeper for `purger.absorb`.

The `scripts` directory is reorganized as a package, and is a workspace 
member of the top-level `opus` package. Additionally, the `scripts` directory 
is itself a workspace containing `deployment` and `simulation` packages 
scripts as its members.