feat: improve trove events (#584)
This PR removes the `Enter` and `Exit` events from Gate, 
adds `Deposit` and `Withdraw` events to the Abbot. and 
adds an `assets` member to the `Shut` event in Caretaker.

As a consequence of these changes, the trove ID no longer 
needs to be passed to the `enter` and `exit` functions in both
Sentinel and Gate.