chore: bump cairo to v2.11.4 (#622)
This commit bumps the Cairo version to `v2.11.4`.

Outdated contracts and scripts have been removed for ease of maintainability. 
To facilitate tracing, versioning has been added to the README for both the 
core contracts and the scripts directory.

Additional changes
- Suggestions by `scarb lint` were implemented e.g. using `for` loops when 
iterating  over spans and `while` loops for conditional loops. This change 
uncovered that the `test_partial_absorb_with_redistribution_entire_trove_debt` 
test did not iterate fully over each nested loop during parametrization. As a 
result, some assertions in the tests had to be relaxed because the edge cases 
were not actually executed previously.
- Replace the use of hash with `generate_arg` to generate price feed for yangs. 
This introduces some randomness to the tests for each CI Run.